support access to internal bdev through fuse node
Add a fuse node for each bdev, appearing as a regular file.

Trouble is that if we make the node appear as a block device to the
application, fuse additionally assumes that to mean to let the kernel
interpret the dev_t as referring to a kernel major/minor, instead of
presenting system calls to our handlers as for other st_mode values.

set -o sync and -o max_readahead=0 options on fuse filesystem

Do not specify direct_io or nonseekable for nodes representing bdevs.